use single class to detect each log framework
